requirements_python313.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
ultralytics==8.0.196
Pillow>=11.0.0
numpy>=1.24.0
python-dotenv==1.0.0
Werkzeug==2.3.7
eventlet==0.33.3 